float16_t

+++ src/my_thing.hpp
@@ -504,11 +504,11 @@ using Thing = struct Thing {
   //
   // Increases per tick and when it reaches 1, allows the thing to move
   //
-  float thing_dt;
+  float16_t thing_dt;
   //
   // Angle of movement.
   //
-  float angle;
+  float16_t angle;
   //
   // Previous map co-ords. Does not change when the move finishes.
   //
